requirements.txt
Django~=3.2.10
django-cors-headers~=3.11.0
djangorestframework~=3.13.1
psycopg2~=2.9.3
